requirements/prod.txt
Django==3.1.5
djangorestframework==3.12.2
python-gnupg==0.4.6
